mdev: separate data retrieval from checking
Introducing a new class for the sysfs data retrieval allows the callers
to make the decision what is going to happen with the retrieved data,
how to handle encountered errors and what checking needs happen before
storing the data into a mdev object.

Adding the methods set_sysfs_data for storing and is_sysfs_data_matching
for checking to the mdev class for easy usability.

Signed-off-by: Boris Fiuczynski <fiuczy@linux.ibm.com>