Remove MDevSysfsData::active
The MDevSysfsData struct contains information about an mdev device that
is looked up from sysfs. But if a mdev is not active, there will be no
information in sysfs. Previously we handled this situation by having a
boolean `active` variable within the MDevSysfsData struct and by making
the various fields within MDevSysfsData have `Option` types so that they
could be `None` when the mdev is not active. To avoid dealing with the
potential for inconsistent data (e.g. `active` set to true, but `parent`
set to `None`), remove the `active` field and change the other fields to
not be `Option`s. Then change the functions that create `MDevSysfsData`
objects to return `Option<MDevSysfsData>`.

For example, previously when calling `MDevSysfsData::load_with_mdev()`
for an inactive mdev, we were returning:
    MDevSysfsData {
      uuid,
      active: false,
      parent: None,
      mdev_type: None,
    }

Now, in the same scenario, we are simply returning None.