Environment doesn't need to be a trait
We had Environment as a trait that was implemented by DefaultEnvironment
(for the cli command) and TestEnvironment (for the test cases). But
this is completely unnecessary as the implementation really only
differs based on the root path. This can just be a common struct with
a configurable root path. The implementation of the TestEnvironment
changes a little bit since it is now a struct that owns an Environment
rather than a struct that implement the Environment trait. So it has
been renamed to TestCase.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>