Change return value of MDevSysfsData::load()
Return a Result<MDevSysfsData> instead of Result<Option<MDevSysfsData>>.
The latter is awkward and doesn't convey any additional information. And
now that we have an error type to represent the fact that a device was
not found, callers can distinguish that situation easily and make their
own decisions about how to handle that case.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>